refactor: move resolveSuiteConfig to harness/config.ts
